+++ enrolsuspension/lib.php
@@ -15,7 +15,7 @@
 // along with Moodle. If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * Version information for the Enrol Suspension plugin.
+ * Navigation hooks for tool_enrolsuspension.
  *
  * @package    tool_enrolsuspension
  * @copyright  2026
@@ -24,8 +24,23 @@
 
 defined('MOODLE_INTERNAL') || die();
 
-$plugin->component = 'tool_enrolsuspension';
-$plugin->version   = 2026081701;
-$plugin->requires  = 2024042200;
-$plugin->release   = '1.1.0-beta';
-$plugin->maturity  = MATURITY_BETA;
+/**
+ * Add a suspension shortcut to course navigation for global suspension operators.
+ *
+ * @param navigation_node $navigation Course navigation node.
+ * @param stdClass $course Course record.
+ * @param context_course $context Course context.
+ */
+function tool_enrolsuspension_extend_navigation_course($navigation, $course, $context): void {
+    if (!has_capability('tool/enrolsuspension:suspend', context_system::instance())) {
+        return;
+    }
+    $url = new moodle_url('/admin/tool/enrolsuspension/index.php', ['courseid' => $course->id]);
+    $navigation->add(
+        get_string('coursesuspensions', 'tool_enrolsuspension'),
+        $url,
+        navigation_node::TYPE_SETTING,
+        null,
+        'tool_enrolsuspension'
+    );
+}